Se realiza template del sidenav y se organiza el modal de los pasos de registro de usuario

+++ qs/src/main/java/com/qs/dao/qs/usuarioMapper.java
@@ -6,7 +6,7 @@ import java.util.List;
 import org.apache.ibatis.annotations.Param;
 
 public interface usuarioMapper {
-    int countByExample(usuarioExample example);
+    long countByExample(usuarioExample example);
 
     int deleteByExample(usuarioExample example);
